refactor(oracle): Extract and compile generator
